+++ src/ads_api/__init__.py
@@ -22,4 +22,4 @@ __all__ = [
     "TokenCredentials",
     "TokenManager",
 ]
-__version__ = "0.9.2"
+__version__ = "0.9.3"